+++ acme/tls_sni_challenge.go
@@ -5,19 +5,16 @@ import (
 	"crypto/sha256"
 	"crypto/tls"
 	"encoding/hex"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"net"
 	"net/http"
-	"time"
 )
 
 type tlsSNIChallenge struct {
-	jws     *jws
-	optPort string
-	start   chan net.Listener
-	end     chan error
+	jws      *jws
+	validate validateFunc
+	iface    string
+	port     string
 }
 
 func (t *tlsSNIChallenge) Solve(chlng challenge, domain string) error {
@@ -26,80 +23,40 @@ func (t *tlsSNIChallenge) Solve(chlng challenge, domain string) error {
 
 	logf("[INFO][%s] acme: Trying to solve TLS-SNI-01", domain)
 
-	t.start = make(chan net.Listener)
-	t.end = make(chan error)
-
 	// Generate the Key Authorization for the challenge
 	keyAuth, err := getKeyAuthorization(chlng.Token, &t.jws.privKey.PublicKey)
 	if err != nil {
 		return err
 	}
 
-	certificate, err := t.generateCertificate(keyAuth)
+	cert, err := t.generateCertificate(keyAuth)
 	if err != nil {
 		return err
 	}
 
-	go t.startSNITLSServer(certificate)
-	var listener net.Listener
-	select {
-	case listener = <-t.start:
-		break
-	case err := <-t.end:
+	// Allow for CLI port override
+	port := "443"
+	if t.port != "" {
+		port = t.port
+	}
+
+	iface := ""
+	if t.iface != "" {
+		iface = t.iface
+	}
+
+	tlsConf := new(tls.Config)
+	tlsConf.Certificates = []tls.Certificate{cert}
+
+	listener, err := tls.Listen("tcp", net.JoinHostPort(iface, port), tlsConf)
+	if err != nil {
 		return fmt.Errorf("Could not start HTTPS server for challenge -> %v", err)
 	}
+	defer listener.Close()
 
-	// Make sure we properly close the HTTP server before we return
-	defer func() {
-		listener.Close()
-		err = <-t.end
-		close(t.start)
-		close(t.end)
-	}()
+	go http.Serve(listener, nil)
 
-	jsonBytes, err := json.Marshal(challenge{Resource: "challenge", Type: chlng.Type, Token: chlng.Token, KeyAuthorization: keyAuth})
-	if err != nil {
-		return errors.New("Failed to marshal network message...")
-	}
-
-	// Tell the server we handle TLS-SNI-01
-	resp, err := t.jws.post(chlng.URI, jsonBytes)
-	if err != nil {
-		return fmt.Errorf("Failed to post JWS message. -> %v", err)
-	}
-
-	// After the path is sent, the ACME server will access our server.
-	// Repeatedly check the server for an updated status on our request.
-	var challengeResponse challenge
-Loop:
-	for {
-		if resp.StatusCode >= http.StatusBadRequest {
-			return handleHTTPError(resp)
-		}
-
-		err = json.NewDecoder(resp.Body).Decode(&challengeResponse)
-		resp.Body.Close()
-		if err != nil {
-			return err
-		}
-
-		switch challengeResponse.Status {
-		case "valid":
-			logf("[INFO][%s] The server validated our request", domain)
-			break Loop
-		case "pending":
-			break
-		case "invalid":
-			return handleChallengeError(challengeResponse)
-		default:
-			return errors.New("The server returned an unexpected state.")
-		}
-
-		time.Sleep(1 * time.Second)
-		resp, err = http.Get(chlng.URI)
-	}
-
-	return nil
+	return t.validate(t.jws, domain, chlng.URI, challenge{Resource: "challenge", Type: chlng.Type, Token: chlng.Token, KeyAuthorization: keyAuth})
 }
 
 func (t *tlsSNIChallenge) generateCertificate(keyAuth string) (tls.Certificate, error) {
@@ -127,27 +84,4 @@ func (t *tlsSNIChallenge) generateCertificate(keyAuth string) (tls.Certificate, 
 	}
 
 	return certificate, nil
-}
-
-func (t *tlsSNIChallenge) startSNITLSServer(cert tls.Certificate) {
-
-	// Allow for CLI port override
-	port := ":443"
-	if t.optPort != "" {
-		port = ":" + t.optPort
-	}
-
-	tlsConf := new(tls.Config)
-	tlsConf.Certificates = []tls.Certificate{cert}
-
-	tlsListener, err := tls.Listen("tcp", port, tlsConf)
-	if err != nil {
-		t.end <- err
-	}
-	// Signal successfull start
-	t.start <- tlsListener
-
-	http.Serve(tlsListener, nil)
-
-	t.end <- nil
 }